add edit function and message

+++ js/script.js
@@ -16,19 +16,30 @@ document.addEventListener("DOMContentLoaded", function () {
         event.preventDefault();
         findBookName();
     });
+    const submitEdit = document.getElementById("editBook");
+    submitEdit.addEventListener("submit", function (event) {
+        console.log("edit initiated");
+        event.preventDefault();
+        updateBookData();
+    });
 
     const revoveAllBtn = document.getElementById("remove-all");
     revoveAllBtn.addEventListener("click", function () {
-        deleteAll();
-        alert("Removed All Data!")
+        if (confirm('Apakah anda ingin menghapus semua Buku?')) {
+            deleteAll();
+        }
     });
     const revoveAllCompleteBtn = document.getElementById("remove-all-complete");
     revoveAllCompleteBtn.addEventListener("click", function () {
+        if (confirm('Apakah anda ingin menghapus semua Completed Buku?')) {
         deleteAllComplete();
+        }
     });
     const revoveAllIncompleteBtn = document.getElementById("remove-all-incomplete");
     revoveAllIncompleteBtn.addEventListener("click", function () {
+        if (confirm('Apakah anda ingin menghapus semua Incomplete Buku?')) {
         deleteAllIncomplete();
+        }
     });
 
     const cancelEditBtn = document.getElementById("cancelEdit");
